third_party/blink/web_tests/external/wpt/css/cssom/getComputedStyle-pseudo-expected.txt

This is a testharness.js-based test.
[FAIL] Pseudo-elements can use the full range of CSS syntax
  assert_equals: expected "50px" but got "100px"
[FAIL] Unknown pseudo-elements
  assert_true: Should return an empty style for unknown pseudo-elements starting with double-colon expected true got false
[FAIL] Unknown pseudo-element with a known string (ex: marker)
  assert_true: Should return an empty style for :marker expected true got false
Harness: the test ran to completion.
